+++ src/Resources/app/administration/src/module/sw-cms/elements/strix-latest-orders/config/index.js
@@ -1,6 +1,6 @@
-import template from "./sw-cms-el-config-strix-latest-orders.html.twig";
+import template from './sw-cms-el-config-strix-latest-orders.html.twig';
 
-Shopware.Component.register("sw-cms-el-config-strix-latest-orders", {
+Shopware.Component.register('sw-cms-el-config-strix-latest-orders', {
     template,
     props: {
         element: { type: Object, required: true },